# skillhub_CICD_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkillHub API",
        "description": "API REST de la plateforme SkillHub : authentification, formations, modules, inscriptions et messagerie.",
        "contact": {
            "name": "MU202612"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/messages/non-lus": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Nombre de messages non lus",
                "description": "Renvoie le nombre de messages non lus reçus par l'utilisateur connecté.\n\nPratique pour afficher un badge \"X messages\" en haut de l'app.",
                "operationId": "537179891317f673b695e97a27066f2d",
                "responses": {
                    "200": {
                        "description": "Compteur renvoyé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/conversations": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Liste des conversations",
                "description": "Renvoie la liste des conversations de l'utilisateur (une ligne par interlocuteur).\n\nÉtapes :\n- Récupère tous les messages où le user est expéditeur OU destinataire,\n  avec leurs relations expediteur et destinataire, triés du plus récent au plus ancien.\n- Itère sur les messages et regroupe par interlocuteur :\n  le 1er rencontré (donc le plus récent) sert d'aperçu.\n- Compte au passage les messages non lus de cet interlocuteur vers le user.\n\nLe résultat ne contient pas l'historique complet, juste un résumé par discussion.",
                "operationId": "8bc13f1bf3394ef201055b9d2444b959",
                "responses": {
                    "200": {
                        "description": "Conversations triées du plus récent au plus ancien"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/{interlocuteurId}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Historique d'une conversation et marquage en lu",
                "description": "Renvoie l'historique complet entre l'utilisateur connecté et un interlocuteur.\n\nEffets de bord : passe à lu=true tous les messages reçus de cet interlocuteur.\nSans ça, le badge \"non lus\" ne baisserait jamais.\n\nLes messages sont triés du plus ancien au plus récent (ordre de lecture).",
                "operationId": "4d9c1d6cc4ed589f0f154ff2a4d9b70c",
                "parameters": [
                    {
                        "name": "interlocuteurId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages renvoyés et marqués lus"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Envoyer un message",
                "description": "Envoie un message à un autre utilisateur.\n\nÉtapes :\n- Vérifie le JWT.\n- Valide destinataire_id (doit exister) et contenu (max 2000 caractères).\n- Détermine si c'est le premier échange entre les deux (aucun message dans aucun sens).\n- Crée le message avec lu=false.\n- Recharge les relations expediteur/destinataire (juste id et nom) pour la réponse.\n- Si premier message, déclenche un mail à l'autre utilisateur via NouveauMessageMail.",
                "operationId": "5157393e27dfe17ac964293ea7dd578e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "destinataire_id",
                                    "contenu"
                                ],
                                "properties": {
                                    "destinataire_id": {
                                        "type": "integer"
                                    },
                                    "contenu": {
                                        "type": "string",
                                        "maxLength": 2000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message créé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/interlocuteurs": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Liste des utilisateurs autorisés à recevoir un message",
                "description": "Renvoie les utilisateurs avec qui l'utilisateur connecté a le droit d'échanger.\n\nRègle métier :\n- Si formateur : tous les apprenants inscrits à au moins une de ses formations.\n- Si apprenant : tous les formateurs des formations auxquelles il est inscrit.\n\nÉvite de pouvoir messager n'importe qui (anti-spam, anti-harcèlement).",
                "operationId": "71ad404246a866ab6f393d4d0995ed61",
                "responses": {
                    "200": {
                        "description": "Liste filtrée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT renvoyé par /api/login. Préfixer avec 'Bearer '.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}